+++ kifi-backend/modules/eliza/app/com/keepit/eliza/commanders/MessageThreadNotificationBuilder.scala
@@ -4,13 +4,18 @@ import com.google.inject.{ Inject, ImplementedBy, Singleton }
 import com.keepit.common.crypto.{ PublicIdConfiguration, PublicId }
 import com.keepit.common.db.Id
 import com.keepit.common.db.slick.DBSession.RSession
+import com.keepit.common.db.slick.Database
+import com.keepit.common.time._
 import com.keepit.discussion.Message
 import com.keepit.eliza.model._
 import com.keepit.model.{ User, NotificationCategory, DeepLocator, Keep }
+import com.keepit.shoebox.ShoeboxServiceClient
 import com.keepit.social.BasicUserLikeEntity
 import org.joda.time.DateTime
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
+
+import scala.concurrent.{ Future, ExecutionContext }
 
 // This is the thing we send to clients so they can display the little
 // "notification" card that summarizes a message thread
@@ -82,17 +87,94 @@ object MessageThreadNotification {
 
 @ImplementedBy(classOf[MessageThreadNotificationBuilderImpl])
 trait MessageThreadNotificationBuilder {
-  def buildForKeeps(userId: Id[User], keepIds: Set[Id[Keep]])(implicit session: RSession): Map[Id[Keep], MessageThreadNotification]
-  def buildForUsers(keepId: Id[Keep], userIds: Set[Id[User]])(implicit session: RSession): Map[Id[User], MessageThreadNotification]
+  def buildForKeeps(userId: Id[User], keepIds: Set[Id[Keep]]): Future[Map[Id[Keep], MessageThreadNotification]]
+  def buildForUsers(keepId: Id[Keep], userIds: Set[Id[User]]): Future[Map[Id[User], MessageThreadNotification]]
+}
+
+object MessageThreadNotificationBuilder {
+  object PrecomputedInfo {
+    case class BuildForKeeps(dummy: Int)
+    case class BuildForUsers(dummy: Int)
+  }
 }
 
 @Singleton
 class MessageThreadNotificationBuilderImpl @Inject() (
+  db: Database,
   messageThreadRepo: MessageThreadRepo,
   userThreadRepo: UserThreadRepo,
-  messageRepo: MessageRepo)
+  messageRepo: MessageRepo,
+  shoebox: ShoeboxServiceClient,
+  implicit val publicIdConfig: PublicIdConfiguration,
+  implicit val executionContext: ExecutionContext)
     extends MessageThreadNotificationBuilder {
-  def buildForKeeps(userId: Id[User], keepIds: Set[Id[Keep]])(implicit session: RSession): Map[Id[Keep], MessageThreadNotification] = ???
-  def buildForUsers(keepId: Id[Keep], userIds: Set[Id[User]])(implicit session: RSession): Map[Id[User], MessageThreadNotification] = ???
+
+  def buildForKeeps(userId: Id[User], keepIds: Set[Id[Keep]]): Future[Map[Id[Keep], MessageThreadNotification]] = {
+    val infoFut = db.readOnlyMasterAsync { implicit s =>
+      val threadsById = messageThreadRepo.getByKeepIds(keepIds)
+      val lastMsgById = keepIds.map { keepId => keepId -> messageRepo.getLatest(keepId) }.toMap
+      val mutedById = keepIds.map { keepId => keepId -> userThreadRepo.isMuted(userId, keepId) }.toMap
+      val threadActivityById = keepIds.map { keepId =>
+        keepId -> userThreadRepo.getThreadActivity(keepId).sortBy { uta =>
+          (-uta.lastActive.getOrElse(START_OF_TIME).getMillis, uta.id.id)
+        }
+      }.toMap
+      val msgCountsById = threadActivityById.map {
+        case (keepId, activity) =>
+          val lastSeenOpt = activity.find(_.userId == userId).flatMap(_.lastSeen)
+          keepId -> messageRepo.getMessageCounts(keepId, lastSeenOpt)
+      }
+      (threadsById, lastMsgById, mutedById, threadActivityById, msgCountsById)
+    }
+    for {
+      (threadsById, lastMsgById, mutedById, threadActivityById, msgCountsById) <- infoFut
+      allUsers = threadsById.values.flatMap(_.allParticipants).toSet
+      basicUserByIdMap <- shoebox.getBasicUsers(allUsers.toSeq)
+    } yield lastMsgById.collect {
+      case (keepId, Some(message)) =>
+        val thread = threadsById(keepId)
+        val threadActivity = threadActivityById(keepId)
+        val (numMessages, numUnread) = msgCountsById(keepId)
+        val muted = mutedById(keepId)
+
+        def basicUserById(id: Id[User]) = basicUserByIdMap.getOrElse(id, throw new Exception(s"Could not get basic user data for $id in MessageThread ${thread.id.get}"))
+        val basicNonUserParticipants = thread.participants.nonUserParticipants.keySet.map(nup => BasicUserLikeEntity(NonUserParticipant.toBasicNonUser(nup)))
+        val messageWithBasicUser = MessageWithBasicUser(
+          id = message.pubId,
+          createdAt = message.createdAt,
+          text = message.messageText,
+          source = message.source,
+          auxData = None,
+          url = message.sentOnUrl.getOrElse(thread.url),
+          nUrl = thread.nUrl,
+          message.from match {
+            case MessageSender.User(id) => Some(BasicUserLikeEntity(basicUserById(id)))
+            case MessageSender.NonUser(nup) => Some(BasicUserLikeEntity(NonUserParticipant.toBasicNonUser(nup)))
+            case _ => None
+          },
+          thread.allParticipants.toSeq.map(u => BasicUserLikeEntity(basicUserById(u))) ++ basicNonUserParticipants.toSeq
+        )
+        val authorActivityInfos = threadActivity.filter(_.lastActive.isDefined)
+
+        val lastSeenOpt: Option[DateTime] = threadActivity.find(_.userId == userId).flatMap(_.lastSeen)
+        val unseenAuthors: Int = lastSeenOpt match {
+          case Some(lastSeen) => authorActivityInfos.count(_.lastActive.get.isAfter(lastSeen))
+          case None => authorActivityInfos.length
+        }
+        keepId -> MessageThreadNotification(
+          message = message,
+          thread = thread,
+          messageWithBasicUser = messageWithBasicUser,
+          unread = !message.from.asUser.contains(userId),
+          originalAuthorIdx = 0,
+          numUnseenAuthors = unseenAuthors,
+          numAuthors = authorActivityInfos.length,
+          numMessages = numMessages,
+          numUnread = numUnread,
+          muted = muted)
+    }
+  }
+
+  def buildForUsers(keepId: Id[Keep], userIds: Set[Id[User]]): Future[Map[Id[User], MessageThreadNotification]] = ???
 }
