+++ kifi-backend/modules/eliza/app/com/keepit/eliza/commanders/NotificationDeliveryCommander.scala
@@ -42,6 +42,7 @@ class NotificationDeliveryCommander @Inject() (
     userThreadRepo: UserThreadRepo,
     nonUserThreadRepo: NonUserThreadRepo,
     messageRepo: MessageRepo,
+    threadNotifBuilder: MessageThreadNotificationBuilder,
     db: Database,
     notificationRouter: WebSocketRouter,
     shoebox: ShoeboxServiceClient,
@@ -209,73 +210,10 @@ class NotificationDeliveryCommander @Inject() (
 
   def buildNotificationForMessageThread(userId: Id[User], keep: Id[Keep]): Future[Option[MessageThreadNotification]] = {
     buildNotificationForMessageThreads(userId, Set(keep)).map(_.get(keep))
-
   }
-  def buildNotificationForMessageThreads(userId: Id[User], keeps: Set[Id[Keep]]): Future[Map[Id[Keep], MessageThreadNotification]] = {
-    val infoFut = db.readOnlyMasterAsync { implicit s =>
-      val threadsById = threadRepo.getByKeepIds(keeps)
-      val lastMsgById = keeps.map { keepId => keepId -> messageRepo.getLatest(keepId) }.toMap
-      val mutedById = keeps.map { keepId => keepId -> userThreadRepo.isMuted(userId, keepId) }.toMap
-      val threadActivityById = keeps.map { keepId =>
-        keepId -> userThreadRepo.getThreadActivity(keepId).sortBy { uta =>
-          (-uta.lastActive.getOrElse(START_OF_TIME).getMillis, uta.id.id)
-        }
-      }.toMap
-      val msgCountsById = threadActivityById.map {
-        case (keepId, activity) =>
-          val lastSeenOpt = activity.find(_.userId == userId).flatMap(_.lastSeen)
-          keepId -> messageRepo.getMessageCounts(keepId, lastSeenOpt)
-      }
-      (threadsById, lastMsgById, mutedById, threadActivityById, msgCountsById)
-    }
-    for {
-      (threadsById, lastMsgById, mutedById, threadActivityById, msgCountsById) <- infoFut
-      allUsers = threadsById.values.flatMap(_.allParticipants).toSet
-      basicUserByIdMap <- shoebox.getBasicUsers(allUsers.toSeq)
-    } yield lastMsgById.collect {
-      case (keepId, Some(message)) =>
-        val thread = threadsById(keepId)
-        val threadActivity = threadActivityById(keepId)
-        val (numMessages, numUnread) = msgCountsById(keepId)
-        val muted = mutedById(keepId)
+  def buildNotificationForMessageThreads(userId: Id[User], keeps: Set[Id[Keep]]): Future[Map[Id[Keep], MessageThreadNotification]] =
+    threadNotifBuilder.buildForKeeps(userId, keeps)
 
-        def basicUserById(id: Id[User]) = basicUserByIdMap.getOrElse(id, throw new Exception(s"Could not get basic user data for $id in MessageThread ${thread.id.get}"))
-        val basicNonUserParticipants = thread.participants.nonUserParticipants.keySet.map(nup => BasicUserLikeEntity(NonUserParticipant.toBasicNonUser(nup)))
-        val messageWithBasicUser = MessageWithBasicUser(
-          id = message.pubId,
-          createdAt = message.createdAt,
-          text = message.messageText,
-          source = message.source,
-          auxData = None,
-          url = message.sentOnUrl.getOrElse(thread.url),
-          nUrl = thread.nUrl,
-          message.from match {
-            case MessageSender.User(id) => Some(BasicUserLikeEntity(basicUserById(id)))
-            case MessageSender.NonUser(nup) => Some(BasicUserLikeEntity(NonUserParticipant.toBasicNonUser(nup)))
-            case _ => None
-          },
-          thread.allParticipants.toSeq.map(u => BasicUserLikeEntity(basicUserById(u))) ++ basicNonUserParticipants.toSeq
-        )
-        val authorActivityInfos = threadActivity.filter(_.lastActive.isDefined)
-
-        val lastSeenOpt: Option[DateTime] = threadActivity.find(_.userId == userId).flatMap(_.lastSeen)
-        val unseenAuthors: Int = lastSeenOpt match {
-          case Some(lastSeen) => authorActivityInfos.count(_.lastActive.get.isAfter(lastSeen))
-          case None => authorActivityInfos.length
-        }
-        keepId -> MessageThreadNotification(
-          message = message,
-          thread = thread,
-          messageWithBasicUser = messageWithBasicUser,
-          unread = !message.from.asUser.contains(userId),
-          originalAuthorIdx = 0,
-          numUnseenAuthors = unseenAuthors,
-          numAuthors = authorActivityInfos.length,
-          numMessages = numMessages,
-          numUnread = numUnread,
-          muted = muted)
-    }
-  }
   def getNotificationsByUser(userId: Id[User], utq: UserThreadQuery, includeUriSummary: Boolean): Future[Seq[NotificationJson]] = {
     val (uts, mts) = db.readOnlyReplica { implicit session =>
       val uts = userThreadRepo.getThreadsForUser(userId, utq)
